# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend REST API for BTMI Market - Seller Platform",
    "title": "BTMI Market API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/buyer/orders": {
      "post": {
        "description": "Create a buyer order, optionally redeeming points. When use_points=true, points are atomically reserved inside the same database transaction as stock reservation (available -> reserved -> consumed on completion; released on reject/cancel). Passing idempotency_key prevents duplicate submissions.",
        "tags": [
          "Buyer Orders"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BuyerCreateOrderRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Buyer order created",
            "schema": {
              "$ref": "#/definitions/BuyerOrderWithLinesResponse"
            }
          },
          "400": {
            "description": "Invalid request, inactive shop, or insufficient points",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Buyer profile, shop, variant or inventory not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Insufficient stock or duplicate order for the same idempotency_key",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "get": {
        "description": "List the authenticated buyer's orders, newest first. Each order includes point reservation fields (base_total, points_used, points_discount_amount, final_total).",
        "tags": [
          "Buyer Orders"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Buyer orders retrieved",
            "schema": {
              "$ref": "#/definitions/BuyerOrdersResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/orders/{order_id}/items/{item_id}/qr/image": {
      "get": {
        "description": "Print-ready PNG of the ORDER_ITEM QR label (seller side). The encoded payload is only the signed opaque tbk.oi.* reference.",
        "tags": [
          "Order Item QR"
        ],
        "produces": [
          "image/png"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "QR label PNG"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - the seller does not own the order's shop"
          },
          "404": {
            "description": "Order or item not found"
          },
          "422": {
            "description": "Unprocessable entity - invalid order or item identifier"
          }
        }
      }
    },
...
    "/buyer/orders/{order_id}/items/{item_id}/qr/image": {
      "get": {
        "description": "Print-ready PNG of the ORDER_ITEM QR label (buyer side). The encoded payload is only the signed opaque tbk.oi.* reference.",
        "tags": [
          "Order Item QR"
        ],
        "produces": [
          "image/png"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "QR label PNG"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - the order does not belong to the buyer"
          },
          "404": {
            "description": "Order or item not found"
          },
          "422": {
            "description": "Unprocessable entity - invalid order or item identifier"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderLineInput": {
      "type": "object",
      "required": [
        "product_id",
        "variant_id",
        "quantity"
      ],
      "properties": {
        "product_id": {
          "type": "string",
          "format": "uuid"
        },
        "variant_id": {
          "type": "string",
          "format": "uuid"
        },
        "quantity": {
          "type": "integer",
          "minimum": 1
        }
      }
    },
...
    "OrderStatusHistoryData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "order_id": {
          "type": "string",
          "format": "uuid"
        },
        "status": {
          "type": "string"
        },
        "changed_by": {
          "type": "string",
          "format": "uuid",
          "nullable": true
        },
        "actor_type": {
          "type": "string",
          "enum": [
            "SELLER",
            "BUYER",
            "SYSTEM"
          ],
          "description": "Who performed the transition"
        },
        "notes": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "BuyerCreateOrderRequest": {
      "type": "object",
      "required": [
        "shop_id",
        "items"
      ],
      "properties": {
        "shop_id": {
          "type": "string",
          "format": "uuid"
        },
        "items": {
          "type": "array",
          "minItems": 1,
          "items": {
            "$ref": "#/definitions/OrderLineInput"
          }
        },
        "use_points": {
          "type": "boolean",
          "default": false,
          "description": "Redeem points on this order (max 20% of the order total)"
        },
        "idempotency_key": {
          "type": "string",
          "description": "Optional unique key. Re-submitting the same key returns DUPLICATE_ORDER (409) instead of creating a second order."
        }
      }
    },
    "BuyerOrderData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "business_id": {
          "type": "string",
          "format": "uuid"
        },
        "shop_id": {
          "type": "string",
          "format": "uuid"
        },
        "customer_id": {
          "type": "string",
          "format": "uuid",
          "nullable": true
        },
        "buyer_profile_id": {
          "type": "string",
          "format": "uuid",
          "nullable": true
        },
        "status": {
          "type": "string",
          "enum": [
            "PENDING",
            "ACCEPTED",
            "REJECTED",
            "PREPARING",
            "COMPLETED",
            "CANCELLED"
          ]
        },
        "total_items": {
          "type": "integer"
        },
        "notes": {
          "type": "string"
        },
        "created_by": {
          "type": "string",
          "format": "uuid",
          "nullable": true
        },
        "base_total": {
          "type": "number",
          "format": "float",
          "description": "Subtotal before any points discount"
        },
        "points_used": {
          "type": "integer",
          "description": "Points reserved for this order (consumed on completion, released on reject/cancel)"
        },
        "points_discount_amount": {
          "type": "number",
          "format": "float",
          "description": "Monetary discount from redeemed points"
        },
        "final_total": {
          "type": "number",
          "format": "float",
          "description": "Total after points discount"
        },
        "idempotency_key": {
          "type": "string",
          "nullable": true
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "delivery_method": {
          "type": "string",
          "description": "PICKUP, SHOP_DELIVERY or PARTNER"
        },
        "delivery_fee_base": {
          "format": "float",
          "description": "Base delivery fee from shop config",
          "type": "number"
        },
        "delivery_points_used": {
          "type": "integer",
          "description": "Delivery points reserved"
        },
        "delivery_points_discount": {
          "format": "float",
          "description": "Delivery discount from points",
          "type": "number"
        },
        "delivery_fee_final": {
          "format": "float",
          "description": "Final delivery fee after point discount",
          "type": "number"
        },
        "delivery_contact_name": {
          "type": "string"
        },
        "delivery_phone": {
          "type": "string"
        },
        "delivery_address": {
          "type": "string"
        },
        "delivery_notes": {
          "type": "string"
        },
        "points_finalized": {
          "type": "boolean",
          "description": "True after the worker consumes reserved points on payment verification"
        }
      }
    },
    "BuyerOrderLineData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "order_id": {
          "type": "string",
          "format": "uuid"
        },
        "product_id": {
          "type": "string",
          "format": "uuid"
        },
        "variant_id": {
          "type": "string",
          "format": "uuid"
        },
        "quantity": {
          "type": "integer"
        },
        "unit_price": {
          "type": "number",
          "format": "float"
        },
        "base_unit_price": {
          "type": "number",
          "format": "float",
          "description": "Price snapshot at order creation before any discount"
        },
        "points_discount_per_unit": {
          "type": "number",
          "format": "float",
          "description": "Points discount applied per unit"
        },
        "final_unit_price": {
          "type": "number",
          "format": "float",
          "description": "Unit price after points discount"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "BuyerOrderWithLinesResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "properties": {
            "order": {
              "$ref": "#/definitions/BuyerOrderData"
            },
            "lines": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BuyerOrderLineData"
              }
            },
            "history": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderStatusHistoryData"
              }
            }
          }
        }
      }
    },
    "BuyerOrdersResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BuyerOrderData"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Enter 'Bearer {token}' - obtain token from POST /auth/login",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}